requirements.txt
argh==0.25.0
beautifulsoup4==4.4.1
chardet==2.3.0
colorlog==2.6.0
Django==1.8.3
django-admin-sortable2==0.3.1
django-appconf==0.6
django-braces==1.4.0
django-classy-tags==0.5.1
django-compressor==1.4
django-extensions==1.3.8
django-filer==0.9.11
django-lineage==0.2.0
django-markdown==0.8.4
django-mptt==0.7.4
-e git://github.com/dokterbob/django-newsletter.git@0664a3c8ff3bfbe3682175da4b84bfbfd801d56d#egg=django_newsletter-master
django-polymorphic==0.7.1
django-sass-processor==0.2.4
-e git://github.com/ojii/django-sekizai.git@6ca7e687c511681f6819848464c3d40286fe8dd0#egg=django_sekizai-0.8.2
django-shell-plus==1.1.5
django-taggit==0.12.2
django-taggit-templatetags==0.2.5
django-tastypie==0.12.1
django-templatetag-sugar==1.0
django-webtest==1.7.8
easy-thumbnails==2.2
gunicorn==19.0.0
libsass==0.8.3
Markdown==2.6.2
meld3==1.0.0
pathtools==0.1.2
Pillow==2.5.1
psycopg2==2.5.3
python-dateutil==2.3
python-mimeparse==0.1.4
pytz==2014.4
PyYAML==3.11
six==1.7.3
sorl-thumbnail==12.3
surlex==0.2.0
Unidecode==0.4.18
vobject==0.6.6
waitress==0.8.10
watchdog==0.8.0
WebOb==1.5.1
WebTest==2.0.20
wheel==0.24.0
